run

Main function. Interacts with the user to select the desired variables for the report, loads them, saves them in an external output, runs the verifications, and informs the user about the success of the whole process. Either the `project_path` should be specified, or the `db_path` with all the related items, being `db_name`, `prj_name`, and `scenarios`.

Usage

run(
  project_path = NULL,
  db_path = NULL,
  db_name = NULL,
  prj_name = NULL,
  scenarios = NULL,
  final_year = 2100,
  desired_variables = "All",
  save_output = TRUE,
  file_name = NULL,
  launch_ui = TRUE
)

Arguments

project_path:

full path of the project with the project name. Possible extensions: .dat and .proj.

db_path:

full path of the database.

db_name:

name of the database.

prj_name:

name of the project.

scenarios:

name of the scenarios to be considered.

final_year:

final year of the data. By default = 2100. ATENTION: final_year must be at least 2025.

desired_variables:

desired variables to have in the report. Considered 'All' by default. Otherwise, specify a vector with all the desired options, being population_clean, GDP_MER_clean, GDP_PPP_clean, global_temp_clean, forcing_clean, co2_concentration_clean, co2_emissions_clean, tot_co2_clean, co2_sequestration_clean, ag_demand_clean, land_clean, primary_energy_clean, energy_trade_clean, elec_gen_tech_clean, elec_capacity_tot_clean, elec_capacity_add_clean, se_gen_tech_clean, fe_sector_clean, energy_service_transportation_clean, energy_service_buildings_clean, ag_prices_clean, industry_production_clean, elec_capital_clean, elec_investment_clean, transmission_invest_clean, CCS_invest_clean, resource_investment_clean, nonco2_clean, co2_price_clean.

save_output:

if TRUE, save reporting data in CSV and XLSX formats. If FALSE, do not save data. If equals 'CSV' or 'XLSX', data saved only in the specified format.

file_name:

file path and name of the saved data. Not used if data not saved. By default, saved in the same directory and with the same name than the specified project_path, with 'iamc_report' tag. CSV and XLSX output. In case of specifing the path, do not introduce the extension, it will be automatically added.

launch_ui:

if TRUE, launch UI, Do not launch UI otherwise.

Value

saved? CSV and XLSX datafile with the desired variables & launched? user interface.